+++ core/jenkins/jenkins_handler.py
@@ -96,9 +96,9 @@ class JenkinsHandler:
             "DEPLOYMENT_SITE": self.trial_network.deployment_site,
             "TNLCM_CALLBACK": self.tnlcm_callback,
             # OPTIONAL
-            "LIBRARY_URL": SixGLibrarySettings.GITHUB_6G_LIBRARY_HTTPS_URL,
+            "LIBRARY_URL": self.trial_network.github_6g_library_https_url,
             "LIBRARY_BRANCH": self.trial_network.github_6g_library_commit_id,
-            "SITES_URL": SixGSandboxSitesSettings.GITHUB_6G_SANDBOX_SITES_HTTPS_URL,
+            "SITES_URL": self.trial_network.github_6g_sandbox_sites_https_url,
             "SITES_BRANCH": self.trial_network.github_6g_sandbox_sites_commit_id,
             "DEBUG": debug
         }
@@ -207,9 +207,9 @@ class JenkinsHandler:
             "DEPLOYMENT_SITE": self.trial_network.deployment_site,
             "TNLCM_CALLBACK": self.tnlcm_callback,
             # OPTIONAL
-            "LIBRARY_URL": SixGLibrarySettings.GITHUB_6G_LIBRARY_HTTPS_URL,
+            "LIBRARY_URL": self.trial_network.github_6g_library_https_url,
             "LIBRARY_BRANCH": self.trial_network.github_6g_library_commit_id,
-            "SITES_URL": SixGSandboxSitesSettings.GITHUB_6G_SANDBOX_SITES_HTTPS_URL,
+            "SITES_URL": self.trial_network.github_6g_sandbox_sites_https_url,
             "SITES_BRANCH": self.trial_network.github_6g_sandbox_sites_commit_id,
             # "SCRIPTED_DESTROY_COMPONENTS": entities_with_destroy_script
             # "DEBUG": true