+++ core/sixg_sandbox_sites/sixg_sandbox_sites_handler.py
@@ -13,7 +13,8 @@ from core.exceptions.exceptions_handler import CustomSixGSandboxSitesException
 class SixGSandboxSitesHandler():
 
     def __init__(
-        self, 
+        self,
+        https_url: str = None,
         reference_type: str = None, 
         reference_value: str = None,
         directory_path: str = None
@@ -25,7 +26,10 @@ class SixGSandboxSitesHandler():
         :param reference_value: value of the reference (branch name, tag name, commit ID) to switch, ``str``
         :param directory_path: directory path into which the 6G-Sandbox-Sites is to be cloned, ``str``
         """
-        self.github_6g_sandbox_sites_https_url = SixGSandboxSitesSettings.GITHUB_6G_SANDBOX_SITES_HTTPS_URL
+        if https_url:
+            self.github_6g_sandbox_sites_https_url = https_url
+        else:
+            self.github_6g_sandbox_sites_https_url = SixGSandboxSitesSettings.GITHUB_6G_SANDBOX_SITES_HTTPS_URL
         self.github_6g_sandbox_sites_repository_name = SixGSandboxSitesSettings.GITHUB_6G_SANDBOX_SITES_REPOSITORY_NAME
         self.github_6g_sandbox_sites_local_directory = os.path.join(directory_path, self.github_6g_sandbox_sites_repository_name)
         self.github_6g_sandbox_sites_reference_type = reference_type